+++ android/lib/model/src/main/kotlin/net/mullvad/mullvadvpn/lib/model/NotificationAction.kt
@@ -9,6 +9,8 @@ sealed interface NotificationAction {
     sealed interface Tunnel : NotificationAction {
         data object Connect : Tunnel
 
+        data object Reconnect : Tunnel
+
         data object Disconnect : Tunnel
 
         data object Cancel : Tunnel